+++ package.json
@@ -16,7 +16,6 @@
   "devDependencies": {
     "@babel/core": "^7.1.2",
     "@babel/preset-env": "^7.1.0",
-    "@types/es6-shim": "^0.31.37",
     "@types/express": "^4.16.0",
     "@types/hls.js": "^0.10.2",
     "@types/node": "^10.12.0",
